# Improve sampling algorithm for AO/CO table