--- a/src/include/cdb/cdbaocsam.h
+++ b/src/include/cdb/cdbaocsam.h
@@ -416,7 +416,10 @@ extern void aocs_delete_finish(AOCSDeleteDesc desc);
 
 extern AOCSHeaderScanDesc aocs_begin_headerscan(
 		Relation rel, int colno);
-extern bool aocs_get_target_tuple(AOCSScanDesc aoscan, int64 targrow, TupleTableSlot *slot);
+extern bool aocs_get_target_tuple(AOCSScanDesc aoscan,
+								  int64 targrow,
+								  TupleTableSlot *slot,
+								  bool checkVisiOnly);
 extern AOCSWriteColumnDesc aocs_writecol_init(Relation rel, List *newvals, AOCSWriteColumnOperation op);
 extern void aocs_writecol_add(Oid relid, List *newvals, List *constraints, TupleDesc oldDesc);
 extern void aocs_writecol_rewrite(Oid relid, List *newvals, TupleDesc oldDesc);