The PR is from this issue: https://github.com/greenplum-db/gpdb/issues/15655

### Issue

When sampling tuples in ANALYZE an AO/CO table, if the sample row number was pointing to a dead tuple, we count this dead row as the current sample slot, instead of skipping it and continuing to get the next live row for this sample slot like Heap. We need to improve the sampling algorithm to select only live tuples.

The PR is for aoco table code path only. There is another code path (`appendonly_acquire_sample_rows()`) for append-only table, and I will open another PR to handle it.

### Basic algorithm

Let `RowSampler` works in the scope of a "filtered" table which includes only live tuples. e.g. there is a table with 100 tuples including 30 live tuples and 70 dead tuples. `RowSampler` can "see" a table with only 30 tuples.

Use a global cursor (`globalOffset`) to remember the actual position. When `RowSampler` skipped `n` tuples before a selecting, move `globalOffset` forward by `n` live tuples (ignoring all dead tuples in the scanning). When `RowSampler` selected a tuple which is dead, continue to move `globalOffset` till getting a live tuple.

### Peformance improvement

The new algorithm can guarantee that all selected tuples are live, and the algorithm of `RowSampler` works as design regardless of layout of live tuples. However, it causes performance regression because we have to perform table scan to read all the live and dead tuples.

For now we use `aocs_get_target_tuple()` to walk through the table, which is unnecessary because we need just to fetch live tuples. **A visimap scan without table scan** will have better performance. However, it is not easy to implement since the code of visimap scan ties to table scan heavily. I tried for a couple of hours and gave up finally because too much effort is needed. We can consider it for performance improvement in future.

### Performance test

Script:

```
drop table aoco1;
create table aoco1 (a int, b int, c int) with (appendonly=true, orientation=column, compresstype=zlib, compresslevel=3);
insert into aoco1 select i,i,i from generate_series(1,10000000)i;
create index on aoco1(a);
delete from aoco1 where a % 4 = 0; -- generate a table with 3/4 live tuples
delete from aoco1 where (a % 4 = 1) or (a % 4 = 2); -- generate a table with 1/4 live tuples
analyze aoco1;
```

Performance:

Old code:

Full table: 468.623 ms
1/4 live tuples: 120.299 ms

New code:

Full table: 1053.901 ms
1/4 live tuples: 910.012 ms

New code is much slower than old code, because new code reads more than 3.3m tuples (on each seg), but old code reads only 10000 tuples.