@haolinw 2023-11-30
Looks like the performance not so good and understand it does introduce more tuples to be scanned, thinking if adding a GUC for temporarily tuning or so might be helpful ?