@yaowangm 2023-11-30
> Looks like the performance not so good and understand it does introduce more tuples to be scanned, thinking if adding a GUC for temporarily tuning (or enable/disable) might be helpful ?

What concrete behavior is the GUC supposed to tune?

As my understanding, there are only two choice: old approach (for incorrect result) and new approach (for correct result). If the GUC is just to enable/disable new approach, it doesn't matter much. In addition, the GUC will import more complexity since we have to maintain two algorithms for varies of scenarios.

On another hand, the performance is not so bad as estimated. As the experiment showed, we got ~1s for scanning 10,000,000 tuples (full table), and can expect ~100s for 1,000,000,000 tuples. I prefer it's acceptable for ANALYZE even a real table may have wider column length.