@haolinw 2023-11-30
> What concrete behavior is the GUC supposed to tune?

Trying to think if it is possible to be tunable based on dead tuples percentage after discussed w/ @kainwen, i.e. even if there are dead tuples in the table, if there is no impact on plan generation then no need to do full scan on the table ?
See "Full table" case (no dead tuples) seemingly introduces significant performance regression but actually it should not and it supposed to be the same as old code ideally.

Well, it may be much tricky to control that accurately, GUC might be an option to disable the full scan manually if dead tuples doesn't impact much.