@yaowangm 2023-11-30
> > What concrete behavior is the GUC supposed to tune?
> 
> Trying to think if it is possible to be tunable based on dead tuples percentage after discussed w/ @kainwen, i.e. even if there are dead tuples in the table, if there is no impact on plan generation then no need to do full scan on the table ? See "Full table" case (no dead tuples) seemingly introduces significant performance regression but actually it shouldn't and it supposed to be the same as old code ideally.
> 
> Well, it may be much tricky to control that accurately, GUC might be an option to disable the full scan manually if dead tuple doesn't impact much.

As discussed, I added some code to handle the special case of full table: we can simply skip the tuples without checking. The performance result:

Old code:

Full table: 466.106 ms
1/4 live tuples: 120.299 ms
3/4 live tuples: 349.179 ms

New code:

Full table: 468.271 ms
1/4 live tuples: 926.711 ms
3/4 live tuples: 1208.782 ms

As expected, now the performance of two algorithms are almost the same for full table, but the difference is still big for skewed tables.

Please take a look to the code. I am thinking about the GUC and other stuffs.