@yaowangm 2023-12-01
@haolinw As talked, I added a GUC `debug_aocotbl_sampling_notablescan` to disable the new algorithm. When the GUC is on, we will follow another code path to emulate the behavior of old code by skipping the actual fetching of tuples. I also used the data set as you suggested:

```
create table aoco1 (a int, b inet, c inet) with (appendonly=true, orientation=column, compresstype=zlib, compresslevel=3);
insert into aoco1 select i, (select ((i%255)::text || '.' || (i%255)::text || '.' || (i%255)::text || '.' ||
(i%255)::text))::inet, (select ((i%255)::text || '.' || (i%255)::text || '.' || (i%255)::text || '.' ||
(i%255)::text))::inet from generate_series(1,10000000)i;
```

I made two small changes to the dataset:

1. I used a smaller data set with about 10,000,000 tuple which should be enough for testing.
2. I used AOCO table (`orientation=column`) because AO table belongs to another code path.

The performance test result:

Original code:

Full table: 547.444
1/4 live tuples: 139.771
3/4 live tuples: 411.533 ms

GUC off:

Full table: 538.617 ms
1/4 live tuples: 1044.632 ms
3/4 live tuples: 1288.147 ms

GUC on:

Full table: 537.489 ms
1/4 live tuples: 137.241 ms
3/4 live tuples: 405.777 ms

As expected, when the GUC is on, we got almost the result to original code.

I am still thinking about the solution 3 for alternative. Let's talk next time.