@yaowangm 2023-12-05
@haolinw As talked, I hacked code and implemented solution 3 ("Vary Step Length Sampling"). See the code: https://github.com/yaowangm/gpdb/tree/aoco_sample_livetuple_vsl

### Perf test

Here is the perf test result:
```
10,000,000 tuples

GUC on (Original code):

Full table: 537.489 ms
1/4 live tuples: 137.241 ms
3/4 live tuples: 405.777 ms

GUC off (solution 5):

Full table: 538.617 ms
1/4 live tuples: 1044.632 ms
3/4 live tuples: 1288.147 ms

Solution 3:

Full table: 1999.858 ms
1/4 live tuples: 1853.039 ms
3/4 live tuples: 1974.550 ms
```
The perf of solution 3 is even worse than solution 5, but the results with different skewed tables are stable.

Things are different for a lager table:
```
100,000,000 tuples

GUC on (Original code):

Full table: 3190.137 ms
1/4 live tuples: 804.078 ms
3/4 live tuples: 2384.549 ms

GUC off (solution 5):

Full table: 3175.981 ms
1/4 live tuples: 8723.943 ms
3/4 live tuples: 11769.748 ms

Solution 3:

Full table: 6708.967 ms
1/4 live tuples: 5716.156 ms
3/4 live tuples: 6418.918 ms
```
For full table, solution 3 has worse result than solution 5, but for skewed table it's much better. Obviously, it's also worse than original code. Additionally, the results are pretty stable for different skewed tables. 

### Analysis

Solution 3 did fetch only 10000 tuples for full table, or ~40000 tuples for 1/4 full table, and solution 5 did almost entire table scan for each case. So why didn't solution 3 get more significant better result than solution 5?

The answer:

1. Aoco table doesn't support random access. When we try to fetch a tuple via `aoco_acquire_sample_rows()`, gpdb just walks forward from current position without skipping unnecessary tuples. The consequence is, randomly fetching 10000 tuples in the table is actually equivalent to entire table scan. (To be accurate, the logic applies only on same seg, but the consequence is valid for a uniform sampling algorithm which needs to access every seg.)
2. Aoco table scan doesn't support "go backward". When we want to fetch a tuple before current position,  `aocs_rescan()` must be called. It explains why solution 3 has worse result than solution 5 in a small table.

My conclusion is, according to current aoco table fetching, **entire table scan is unavoidable for skewed table**. 

I also considered index only scan for aoco table, but it seems relevant interface (`aoco_index_fetch_tuple()`) doesn't support fetching by arbitrary row number. In addition, index only scan cannot fit all sampling scenarios.

Let's talk later.