@yaowangm 2023-12-07
@haolinw I have implemented visimap scan without table scan by adding a parameter `checkVisiOnly`. Please see the latest code.

Here is the perf test result (for 10m tuples):
```
without checkVisiOnly:

Full table: 538.617 ms
1/4 live tuples: 1044.632 ms
3/4 live tuples: 1288.147 ms

with checkVisiOnly:

Full table: 549.458 ms
1/4 live tuples: 759.397 ms
3/4 live tuples: 760.660 ms

Original code:

Full table: 537.489 ms
1/4 live tuples: 137.241 ms
3/4 live tuples: 405.777 ms
```
For full table, the results are almost the same.

For 3/4 and 1/4 live tuples, the code w/ checkVisiOnly is much better then the code wout/ checkVisiOnly, but still worse than original code.

Comparing the code w/ checkVisiOnly and original code:

For 1/4 live tuples, original code actually fetched 1/4 selected tuples since it didn't check the visibility. If it fetches all selected tuples, the cost should be 137 x 4 = 548ms. For the code w/ checkVisiOnly, the rest (759 - 548 = ) 211ms is for extra visibility checking. 

For 3/4 live tuples, if original code fetches all selected tuples, the cost should be 405 /  (3 / 4) = 540ms. For the code w/ checkVisiOnly, the rest (760 - 540 = ) 200ms is for extra visibility checking. 

Since a correct algorithm must fetch all selected tuples, the regression issue may not be so serious as we thought.

There is still space for perf improvement in the cost of extra visibility checking, such as avoiding to initialize column info for each fetching. But I am not sure whether it is worthy since most of cost belongs to IO.

Let's talk later.