@yaowangm 2023-12-08
As talked, I rebased to the latest code of main branch, and re-ran tests on a large data set about 1.28b tuples.

Here is the script:
```
drop table aoco1;
create table aoco1 (a int, b inet, c inet) with (appendonly=true, orientation=column, compresstype=zlib, compresslevel=3);
insert into aoco1 select i, (select ((i%255)::text || '.' || (i%255)::text || '.' || (i%255)::text || '.' ||
(i%255)::text))::inet, (select ((i%255)::text || '.' || (i%255)::text || '.' || (i%255)::text || '.' ||
(i%255)::text))::inet from generate_series(1,10000000)i;
insert into aoco1 select * from aoco1;
insert into aoco1 select * from aoco1;
insert into aoco1 select * from aoco1;
insert into aoco1 select * from aoco1;
insert into aoco1 select * from aoco1;
insert into aoco1 select * from aoco1;
insert into aoco1 select * from aoco1;
create index on aoco1(a);
delete from aoco1 where a % 4 = 0; -- generate a table with 3/4 live tuples
delete from aoco1 where (a % 4 = 1) or (a % 4 = 2); -- generate a table with 1/4 live tuples
analyze aoco1;
```
The results:
```
Original code: 

Full table: 6332.089 ms
1/4 live tuples: 1627.982 ms
3/4 live tuples: 4818.294 ms

New code:

Full table: 6332.580 ms
1/4 live tuples: 34627.929 ms
3/4 live tuples: 34770.795 ms
```
The results of original code shows the cost for fetching 10000 tuples in arbitrary positions is about 63xx ms.

For 3/4 live tuples, the results of new code shows the cost for visibility checking is about (34770 - 6332 = ) 28438ms.  For 1/4 live tuples, the cost is (34627 - 6332 =) 28295ms. The numbers make sense because the expected number is (220 x 128 =) 28160 since about 220ms is needed for 10m tuples in previous tests.

As talked, we can treat the performance of visimap as a separated issue if necessary.